gen_video/requirements.txt
# Stability AI Generative Models 依赖
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
safetensors>=0.3.0
omegaconf>=2.3.0
einops>=0.6.0
tqdm>=4.65.0
pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0

# InstantID 相关
instantid>=0.1.0
insightface>=0.7.3
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0  # 如果有 GPU，使用 GPU 版本

# CosyVoice 相关（需要从 GitHub 安装）
# git+https://github.com/FunAudioLLM/CosyVoice.git
modelscope>=1.9.0  # 用于下载 CosyVoice 模型

# 视频处理
ffmpeg-python>=0.2.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# 音频处理（TTS）
chattts>=0.3.0
librosa>=0.10.0
soundfile>=0.12.0
torchaudio>=2.0.0

# 字幕生成（WhisperX）
whisperx>=3.1.1
stable-ts>=2.11.0
faster-whisper>=0.10.0
pyannote.audio>=3.1.0

# 其他工具
pydub>=0.25.1
pytube>=15.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
huggingface-hub>=0.19.0


